split some components up and made home page useful
