+++ src/App.js
@@ -37,7 +37,7 @@ export class App extends Component {
           <div>
             <Navigation />
             <Tab.Container
-              defaultActiveKey="2"
+              defaultActiveKey="1"
               className="justify-content-between"
             >
               <Row>
